+++ src/main/java/ro/cucumber/core/clients/http/Method.java
@@ -1,5 +1,5 @@
 package ro.cucumber.core.clients.http;
 
 public enum Method {
-    GET, POST, PUT, DELETE, HEAD, OPTIONS, TRACE
+    GET, POST, PUT, DELETE, HEAD, OPTIONS, TRACE, PATCH
 }