+++ Changelog.md
@@ -1,6 +1,13 @@
+### 2.0.2
+Refactoring  
+Add default data table converter: JacksonTableTransformer  
+Add sslContext, requestRetryHandler and serviceRetryStrategy build methods  
+HttpClient now supports PATCH
+
 ### 2.0.1
-Add logger for http client  
+Add interceptor logger http client  
 Colorize logging
+
 ### 2.0.0
 Use Cucumber 4+  
 Add more helpers for reading resources